+++ src/main/java/jewellery/inventory/exception/not_found/ProductNotFoundException.java
@@ -7,7 +7,4 @@ public class ProductNotFoundException extends NotFoundException{
     public ProductNotFoundException(UUID id) {
         super("The product with id " + id + " is not found");
     }
-    public ProductNotFoundException(String name) {
-        super("The product with name " + name + " is not found");
-    }
 }
