+++ src/main/java/jewellery/inventory/exception/not_found/UserNotFoundException.java
@@ -6,8 +6,4 @@ public class UserNotFoundException extends NotFoundException {
   public UserNotFoundException(UUID id) {
     super("User with id " + id + " was not found");
   }
-
-  public UserNotFoundException(String name) {
-    super("User with name " + name + " was not found");
-  }
 }
