+++ src/test/java/jewellery/inventory/helper/ProductTestHelper.java
@@ -2,6 +2,7 @@ package jewellery.inventory.helper;
 
 import jewellery.inventory.dto.request.ProductRequestDto;
 import jewellery.inventory.dto.request.resource.ResourceQuantityRequestDto;
+import jewellery.inventory.dto.response.ResourcesInUserResponseDto;
 import jewellery.inventory.model.Product;
 import jewellery.inventory.model.ResourceInUser;
 import jewellery.inventory.model.User;
@@ -9,6 +10,7 @@ import jewellery.inventory.model.resource.Resource;
 import jewellery.inventory.model.resource.ResourceInProduct;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -32,6 +34,35 @@ public class ProductTestHelper {
         productRequestDto.setResourcesContent(List.of(resourceQuantityRequestDto));
         productRequestDto.setSalePrice(10000);
         return productRequestDto;
+    }
+
+    @NotNull
+    public static ProductRequestDto getProductRequestDto(ResourcesInUserResponseDto resourcesInUser, User user) {
+
+        List<ResourceQuantityRequestDto> listOfResourcesInProduct = getResourceInProductRequestDtos(resourcesInUser);
+
+        ProductRequestDto productRequestDto = new ProductRequestDto();
+        productRequestDto.setName("TestProduct");
+        productRequestDto.setAuthors(List.of("TestAuthors"));
+        productRequestDto.setDescription("Test");
+        productRequestDto.setOwnerId(user.getId());
+        productRequestDto.setSalePrice(50);
+        productRequestDto.setResourcesContent(listOfResourcesInProduct);
+
+        return productRequestDto;
+    }
+
+    @NotNull
+    private static List<ResourceQuantityRequestDto> getResourceInProductRequestDtos(ResourcesInUserResponseDto resourcesInUser) {
+        ResourceQuantityRequestDto resourceQuantityRequestDto = new ResourceQuantityRequestDto();
+        List<ResourceQuantityRequestDto> listOfResourcesInProduct = new ArrayList<>();
+        resourcesInUser.getResourcesAndQuantities().forEach(r ->
+        {
+            resourceQuantityRequestDto.setId(r.getResource().getId());
+            resourceQuantityRequestDto.setQuantity(5);
+            listOfResourcesInProduct.add(resourceQuantityRequestDto);
+        });
+        return listOfResourcesInProduct;
     }
 
     public static ResourceQuantityRequestDto createResourceQuantityRequestDto(Resource pearl) {
